Create README.md
The WaveICA R package provides a new algorithm to removing batch effects for metabolomics data. The details of this method are in the paper "Removal of batch effects for large-scale untargeted metabolomics data based on wavelet transform". To install this package in R, using this command: devtools::install_github("dengkuistat/WaveICA")